requirements.txt
streamlit>=1.36
opencv-python
numpy
pillow
matplotlib
requests
streamlit-webrtc